# api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "TAMS API",
    "description": "Time-addressable Media Store API - A comprehensive FastAPI implementation of the BBC TAMS API specification with VAST database integration for high-performance time-series analytics and S3-compatible storage for media segments.",
    "version": "6.0",
    "contact": {
      "name": "TAMS API Support",
      "url": "https://github.com/jesseVast/vasttams"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.tams.example.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/flows": {
      "head": {
        "summary": "Return flows path headers",
        "operationId": "head_flows",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "summary": "List flows with optional filtering",
        "operationId": "list_flows",
        "tags": [
          "Flows"
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "query",
            "description": "Filter by source ID",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "timerange",
            "in": "query",
            "description": "Filter by time range",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Filter by format",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "urn:x-nmos:format:video",
                "urn:x-nmos:format:audio",
                "urn:x-nmos:format:data",
                "urn:x-tam:format:image",
                "urn:x-nmos:format:multi"
              ]
            }
          },
          {
            "name": "codec",
            "in": "query",
            "description": "Filter by codec",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "label",
            "in": "query",
            "description": "Filter by label",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "frame_width",
            "in": "query",
            "description": "Filter by frame width",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "frame_height",
            "in": "query",
            "description": "Filter by frame height",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Pagination key",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of flows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new flow",
        "operationId": "create_flow",
        "tags": [
          "Flows"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flow"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Flow created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flow"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/flows/{flow_id}": {
      "head": {
        "summary": "Return flow path headers",
        "operationId": "head_flow",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "summary": "Get flow by ID",
        "operationId": "get_flow",
        "tags": [
          "Flows"
        ],
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flow details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flow"
                }
              }
            }
          },
          "404": {
            "description": "Flow not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update flow",
        "operationId": "update_flow",
        "tags": [
          "Flows"
        ],
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flow"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flow updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flow"
                }
              }
            }
          },
          "404": {
            "description": "Flow not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete flow",
        "operationId": "delete_flow",
        "tags": [
          "Flows"
        ],
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flow deleted successfully"
          },
          "404": {
            "description": "Flow not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Flow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "source_id": {
            "type": "string",
            "format": "uuid"
          },
          "format": {
            "type": "string",
            "enum": [
              "urn:x-nmos:format:video",
              "urn:x-nmos:format:audio",
              "urn:x-nmos:format:data",
              "urn:x-tam:format:image",
              "urn:x-nmos:format:multi"
            ]
          },
          "codec": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "updated_by": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "container": {
            "type": "string"
          },
          "read_only": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "id",
          "source_id",
          "format",
          "codec"
        ]
      },
      "FlowsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Flow"
            }
          },
          "paging": {
            "$ref": "#/components/schemas/PagingInfo"
          }
        }
      },
...
      "PagingInfo": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer"
          },
          "next_key": {
            "type": "string"
          }
        }
      }
    }
  },
...
}